requirements.txt
mlflow==2.1.1
numpy==1.23.5
Pillow==9.4.0
PyYAML==6.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
